Auth: move did:web public endpoints to VDR v2 API
