+++ auth/api/iam/api_test.go
@@ -122,100 +122,6 @@ func TestWrapper_RootOAuthAuthorizationServerMetadata(t *testing.T) {
 	})
 }
 
-func TestWrapper_GetTenantWebDID(t *testing.T) {
-	const webIDPart = "123"
-	ctx := audit.TestContext()
-	expectedWebDIDDoc := did.Document{
-		ID: webDID,
-	}
-	// remarshal expectedWebDIDDoc to make sure in-memory format is the same as the one returned by the API
-	data, _ := json.Marshal(expectedWebDIDDoc)
-	_ = expectedWebDIDDoc.UnmarshalJSON(data)
-
-	t.Run("ok", func(t *testing.T) {
-		test := newTestClient(t)
-		test.vdr.EXPECT().ResolveManaged(webDID).Return(&expectedWebDIDDoc, nil)
-
-		response, err := test.client.GetTenantWebDID(ctx, GetTenantWebDIDRequestObject{webIDPart})
-
-		assert.NoError(t, err)
-		assert.Equal(t, expectedWebDIDDoc, did.Document(response.(GetTenantWebDID200JSONResponse)))
-	})
-	t.Run("non-root base URL", func(t *testing.T) {
-		nonRootWebDID := did.MustParseDID("did:web:example.com:identities:tenant1:iam:123")
-		expectedWebDIDDoc := did.Document{
-			ID: nonRootWebDID,
-		}
-		data, _ := json.Marshal(expectedWebDIDDoc)
-		_ = expectedWebDIDDoc.UnmarshalJSON(data)
-
-		baseURL, _ := url.Parse("https://example.com/identities/tenant1")
-		test := newTestClientWithBaseURL(t, baseURL)
-		test.vdr.EXPECT().ResolveManaged(nonRootWebDID).Return(&expectedWebDIDDoc, nil)
-
-		response, err := test.client.GetTenantWebDID(ctx, GetTenantWebDIDRequestObject{webIDPart})
-
-		assert.NoError(t, err)
-		assert.Equal(t, expectedWebDIDDoc, did.Document(response.(GetTenantWebDID200JSONResponse)))
-	})
-	t.Run("unknown DID", func(t *testing.T) {
-		test := newTestClient(t)
-		test.vdr.EXPECT().ResolveManaged(webDID).Return(nil, resolver.ErrNotFound)
-
-		response, err := test.client.GetTenantWebDID(ctx, GetTenantWebDIDRequestObject{webIDPart})
-
-		assert.NoError(t, err)
-		assert.IsType(t, GetTenantWebDID404Response{}, response)
-	})
-	t.Run("other error", func(t *testing.T) {
-		test := newTestClient(t)
-		test.vdr.EXPECT().ResolveManaged(webDID).Return(nil, errors.New("failed"))
-
-		response, err := test.client.GetTenantWebDID(ctx, GetTenantWebDIDRequestObject{webIDPart})
-
-		assert.EqualError(t, err, "unable to resolve DID")
-		assert.Nil(t, response)
-	})
-}
-
-func TestWrapper_GetRootWebDID(t *testing.T) {
-	ctx := audit.TestContext()
-	expectedWebDIDDoc := did.Document{
-		ID: rootWebDID,
-	}
-	// remarshal expectedWebDIDDoc to make sure in-memory format is the same as the one returned by the API
-	data, _ := json.Marshal(expectedWebDIDDoc)
-	_ = expectedWebDIDDoc.UnmarshalJSON(data)
-
-	t.Run("ok", func(t *testing.T) {
-		test := newTestClient(t)
-		test.vdr.EXPECT().ResolveManaged(rootWebDID).Return(&expectedWebDIDDoc, nil)
-
-		response, err := test.client.GetRootWebDID(ctx, GetRootWebDIDRequestObject{})
-
-		assert.NoError(t, err)
-		assert.Equal(t, expectedWebDIDDoc, did.Document(response.(GetRootWebDID200JSONResponse)))
-	})
-	t.Run("unknown DID", func(t *testing.T) {
-		test := newTestClient(t)
-		test.vdr.EXPECT().ResolveManaged(rootWebDID).Return(nil, resolver.ErrNotFound)
-
-		response, err := test.client.GetRootWebDID(ctx, GetRootWebDIDRequestObject{})
-
-		assert.NoError(t, err)
-		assert.IsType(t, GetRootWebDID404Response{}, response)
-	})
-	t.Run("other error", func(t *testing.T) {
-		test := newTestClient(t)
-		test.vdr.EXPECT().ResolveManaged(rootWebDID).Return(nil, errors.New("failed"))
-
-		response, err := test.client.GetRootWebDID(ctx, GetRootWebDIDRequestObject{})
-
-		assert.EqualError(t, err, "unable to resolve DID")
-		assert.Nil(t, response)
-	})
-}
-
 func TestWrapper_GetOAuthClientMetadata(t *testing.T) {
 	t.Run("ok", func(t *testing.T) {
 		ctx := newTestClient(t)
