+++ docs/_static/vdr/v2.yaml
@@ -375,6 +375,48 @@ paths:
                   $ref: '#/components/schemas/VerificationMethod'
         default:
           $ref: '../common/error_response.yaml'
+  /iam/{id}/did.json:
+    parameters:
+      - name: id
+        in: path
+        description: ID of DID.
+        required: true
+        example: EwVMYK2ugaMvRHUbGFBhuyF423JuNQbtpes35eHhkQic
+        schema:
+          type: string
+    get:
+      summary: Returns the did:web DID for the specified tenant.
+      description: |
+        Returns the did:web DID for the specified tenant, if it is owned by this node.
+      operationId: "getTenantWebDID"
+      tags:
+        - DID
+      responses:
+        "200":
+          description: DID has been found and returned.
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/DIDDocument'
+        "404":
+          description: DID does not exist.
+  /.well-known/did.json:
+    get:
+      summary: Returns the root did:web DID of this domain.
+      description: |
+        Returns the root did:web DID of this domain, if it is owned by this node.
+      operationId: "getRootWebDID"
+      tags:
+        - DID
+      responses:
+        "200":
+          description: DID has been found and returned.
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/DIDDocument'
+        "404":
+          description: DID does not exist.
 #  /internal/vdr/v2/did/{did}/verificationmethod/{id}:
 #    parameters:
 #      - name: did
